feat: add prelude module
