+++ client-grpc/Cargo.toml
@@ -13,16 +13,11 @@ repository.workspace   = true
 
 [features]
 dev       = ["mock", "tokio"]
-test_util = ["mock", "stub_backends"]
+test_util = ["mock", "stub_backends", "tokio"]
 # Will add a 'mock' module for the enabled resources, providing access to mock data generation functions
 mock = []
 # Will use a stubbed server connection, only use for tests!
-stub_backends = [
-  "svc-assets/stub_server",
-  "lib-common/grpc_mock",
-  "tower",
-  "tokio",
-]
+stub_backends = ["svc-assets/stub_server", "lib-common/grpc_mock", "tower"]
 # Will implement stub functions for the client, only use for tests!
 stub_client = ["svc-assets"]
